Fix crash in release mode
The problem (that was especially visible in Release mode) was a crash in the
destructor of the RNSkDrawViewImpl on iOS where the CAMetalLayer was
destructed twice in certain scenarios.

The solution was to rewrite ownership of the RNSkDrawView so that we could
make sure none of these objects were destroyed before all threads were done
drawing etc. Each RNSkDrawView is now stored in a shared_ptr and captures
are kept using weak_ptrs to ensure the lifetime of these objects.

- Removed all std::bind that did not capture ownership correctly
- Removed setNativeDrawFunc and converted to virtual function
- Changed all references to RNSkDrawView to be shared_ptrs
  - Changed captures of RNSkDrawView in lambdas to use weak_ptrs
  - Changed caputers of RNSk***Value in lambdas to use weak_ptrs
- Updated RNSkTimingInfo to handle skipping of frames correctly
   It now displays the time of the frame that made us skip the next frame.
- Updated Android source code
  - Moved non-jni code to directory rnskia-android (alignment with iOS)
  - Fixed cleanup/dtor issues in Android
  - Added RNSkDrawView descendant RNSkDrawViewImpl to be able to store
    reference as std::shared_ptr (alignment with iOS)
- iOS
  - Fixed cleanup issues, made sure CAMetalLayer is destroyed on the main
    UI Thread