RNWeb-preparations: Refactor folders and types (#474)
* Refactored Views folder to prepare for RNWeb

- Moved types to types.ts to split implementation and types
- Moved declaration of SkiaViewApi to types
- Moved implementation (import from global) to api file

* Refactored Values folder to prepare for RNWeb

Moved types to types.ts and implementation in api.ts to separate implementation and declaration.

* Moved animations out of values and on to root in package

This is done to separate the API that has cross platform requirements with implementations that uses these APIs. The animation module uses the Values api but does not itself contain any specific platform code.

* Refactored Views folder to prepare for RNWeb

- Moved types to types.ts to split implementation and types
- Moved declaration of SkiaViewApi to types
- Moved implementation (import from global) to api file

* Refactored Values folder to prepare for RNWeb

Moved types to types.ts and implementation in api.ts to separate implementation and declaration.

* Moved animations out of values and on to root in package

This is done to separate the API that has cross platform requirements with implementations that uses these APIs. The animation module uses the Values api but does not itself contain any specific platform code.