Refactored declaration to use a declaration context
This is an alignment with the JS code to add support for multiple nested composable elements in the DOM in a safe and generic way.

The refactoring introduces a DeclarationContext containing a stack of composables for the ImageFilter, ColorFilter, MaskFilter, Shader and PathEffect nodes so that they can push their results which can finally be used to decorate the paint node in a render node.

I've also removed the use of std::dynamic_pointer_cast to detect node type - it showed up in the profiler. Instead we're now using getDeclarationType and an enum.